added files of version 0.7.0 (Galaxius)

+++ include/bundle7/timestamp.h
@@ -1,11 +1,12 @@
 #ifndef BUNDLE7_TIMESTAMP_H_INCLUDED
 #define BUNDLE7_TIMESTAMP_H_INCLUDED
 
-#include <stdint.h>
 #include "cbor.h"
+
+#include <stdint.h>
 
 CborError bundle7_timestamp_parse(CborValue * it,
 	uint64_t *creation_timestamp,
-	uint16_t *sequence_number);
+	uint64_t *sequence_number);
 
 #endif // BUNDLE7_TIMESTAMP_H_INCLUDED
